split up FilterList
